Split ImageMethods.cs, move the headless bitmap helpers to Core.Windows
Batch 5c of the Core split.

Call-site census before deciding the split direction: PadImage (0 external
callers - only used internally by GetRegionOfScreenAsBitmap, which stays),
GetBitmapFromIRandomAccessStream (3: BarcodeUtilitiesTests, OcrSourceUtilities,
PdfDocumentRenderer) and GetRotateFlipType(string) (2: GrabFrame.xaml.cs,
OcrSourceUtilities) total 5 external call sites. Everything else in the file
- BitmapImageToBitmap, BitmapToImageSource, CachedBitmapToBitmapImage,
GetRegionOfScreenAsBitmap, GetWindowsBoundsBitmap, GetWindowBoundsImage,
ScaleBitmapUniform, InteropBitmapToBitmap, BitmapSourceToBitmap,
ImageSourceToBitmap, GetBitmapImageFromIRandomAccessStream, RotateImage - has
on the order of 60 call sites across GrabFrame.xaml.cs, FullscreenGrab,
MagickHelpers, NotifyIconUtilities, OcrSourceUtilities and PdfDocumentRenderer.
Unlike 4c, the census here points the other way: the portable half is the
minority, so it took the new name (Text-Grab.Core.Windows/Utilities/
BitmapUtilities.cs) and the app-bound majority kept ImageMethods.

CaptureScreenRegion did not move, and is recorded as a new row in section 7.
It calls HdrScreenCapture.TryCaptureRegion when HdrCaptureCorrection is set,
and HdrScreenCapture.cs is explicitly out of this batch's scope (batch 5b,
being done separately, needs a settable dispatcher hook first). Core.Windows
cannot reference the app, so CaptureScreenRegion cannot move until
HdrScreenCapture does. Its only two callers (GetRegionOfScreenAsBitmap,
GetWindowsBoundsBitmap) already stay behind for their own reasons (invariant
5's HistoryService caching, and the GrabFrame View pattern-match), so nothing
else was blocked by leaving it. Because CaptureScreenRegion stayed, its
AppUtilities.TextGrabSettings.HdrCaptureCorrection read stayed with it -
HdrCaptureCorrection was deliberately NOT added to ITextGrabSettings, since
nothing that actually moved reads it.

BitmapUtilities.cs keeps ImageMethods' original namespace (Text_Grab, not
Text_Grab.Utilities - the file was always oddly namespaced) per invariant 1.
GetRotateFlipType(string) still calls the GetRotateFlipType(this Image)
extension from Extensions/ImageExtensions.cs, already in Core.Windows since
5a.

Self-assembly-reflection audit: grepped BitmapUtilities.cs for
GetExecutingAssembly/GetEntryAssembly/GetCallingAssembly, typeof(x).Assembly,
Assembly.GetName/.Location/.CodeBase, AssemblyMetadata, and
AppContext.BaseDirectory - zero matches.

Gate: dotnet build Tests/Tests.csproj -c Debug -p:Platform=x64 - clean
(0 errors; pre-existing MSB3277 assembly-conflict warnings only).
Tests.Core 514, Tests.Core.Windows 3, Tests 993 passed/7 skipped -
identical to baseline, total 1510 conserved.
