+++ Text-Grab/Utilities/ImageMethods.cs
@@ -19,24 +19,6 @@ namespace Text_Grab;
 
 public static class ImageMethods
 {
-    public static Bitmap PadImage(Bitmap image, int minW = 64, int minH = 64)
-    {
-        if (image.Height >= minH && image.Width >= minW)
-            return image;
-
-        int width = Math.Max(image.Width + 16, minW + 16);
-        int height = Math.Max(image.Height + 16, minH + 16);
-
-        // Create a compatible bitmap
-        Bitmap destination = new(width, height, image.PixelFormat);
-        using Graphics gd = Graphics.FromImage(destination);
-
-        gd.Clear(image.GetPixel(0, 0));
-        gd.DrawImageUnscaled(image, 8, 8);
-
-        return destination;
-    }
-
     public static Bitmap BitmapImageToBitmap(BitmapImage bitmapImage)
     {
         using MemoryStream outStream = new();
@@ -116,7 +98,7 @@ public static class ImageMethods
     public static Bitmap GetRegionOfScreenAsBitmap(Rectangle region, bool cacheResult = true)
     {
         Bitmap bmp = CaptureScreenRegion(region);
-        bmp = PadImage(bmp);
+        bmp = BitmapUtilities.PadImage(bmp);
 
         if (cacheResult)
             Singleton<HistoryService>.Instance.CacheLastBitmap(bmp);
@@ -246,16 +228,6 @@ public static class ImageMethods
         };
     }
 
-    public static Bitmap GetBitmapFromIRandomAccessStream(IRandomAccessStream stream)
-    {
-        Stream managedStream = stream.AsStream();
-        if (managedStream.CanSeek)
-            managedStream.Position = 0;
-
-        using Bitmap bitmap = new(managedStream);
-        return new Bitmap(bitmap);
-    }
-
     public static BitmapImage GetBitmapImageFromIRandomAccessStream(IRandomAccessStream stream)
     {
         BitmapImage bmp = new();
@@ -268,13 +240,6 @@ public static class ImageMethods
         bmp.EndInit();
         bmp.Freeze();
         return bmp;
-    }
-
-    internal static RotateFlipType GetRotateFlipType(string path)
-    {
-        using Image img = Image.FromFile(path);
-        RotateFlipType rotateFlipType = img.GetRotateFlipType();
-        return rotateFlipType;
     }
 
     internal static void RotateImage(BitmapImage droppedImage, RotateFlipType rotateFlipType)
