+++ docs/Core-Split-Plan.md
@@ -699,6 +699,7 @@ Every row below was verified by reading the file, not inferred.
 | `Utilities/GrabTemplateExecutor.cs` | `LoadStoredRegexes()` needs a non-scalar seam. `GrabTemplate`/`TemplateRegion` moved to Core in 2d and 4c has landed, so the remaining blocker is the settings façade plus its calls into `OcrSourceUtilities` | a façade + `OcrSourceUtilities` |
 | `Utilities/PdfDocumentRenderer.cs` | `RenderPageAsync` returns `BitmapSource` — a public API shape change affecting several views | 5f |
 | `Services/HistoryService.cs` | Headless JSON pipeline interleaved with WPF menu building and `GrabFrame`/`EditTextWindow` construction, sharing private state | Opus split (6e) |
+| `Utilities/ImageMethods.cs: CaptureScreenRegion` | Calls `HdrScreenCapture.TryCaptureRegion`, which stays in the app pending 5b's settable dispatcher hook. Core.Windows cannot reference the app, so this cannot move until `HdrScreenCapture.cs` does. `PadImage`, `GetBitmapFromIRandomAccessStream` and `GetRotateFlipType(string)` moved on their own to `Text-Grab.Core.Windows/Utilities/BitmapUtilities.cs`; `HdrCaptureCorrection` was **not** added to `ITextGrabSettings` because nothing that actually moved reads it - `CaptureScreenRegion` still reads it via `AppUtilities.TextGrabSettings` in the app | 5b |
 
 ## 8. Verified dead code — free, zero-risk prep
